+++ src/main/java/chess/controller/state/End.java
@@ -2,8 +2,8 @@ package chess.controller.state;
 
 import chess.controller.Command;
 import chess.controller.ScoreDto;
-import chess.domain.Score;
-import chess.domain.chess.ChessGame;
+import chess.domain.game.ChessGame;
+import chess.domain.result.Score;
 import chess.view.OutputView;
 
 public final class End implements State {
