+++ src/test/java/chess/domain/game/ChessGameTest.java
@@ -1,4 +1,4 @@
-package chess.domain.chess;
+package chess.domain.game;
 
 import chess.domain.position.Position;
 import org.junit.jupiter.params.ParameterizedTest;
